+++ scripts/mergelora.py
@@ -11,7 +11,7 @@ from mobilevlm.constants import DEFAULT_IMAGE_PATCH_TOKEN, DEFAULT_IM_START_TOKE
 
 def merge_lora(model_base, model_path, save_path):
 
-    kwargs = {"device_map": "auto", "torch_dtype": torch.float16}
+    kwargs = {"device_map": "auto", "torch_dtype": torch.float32}
 
     lora_cfg_pretrained = AutoConfig.from_pretrained(model_path)
     tokenizer = AutoTokenizer.from_pretrained(model_base, use_fast=False)
@@ -42,7 +42,7 @@ def merge_lora(model_base, model_path, save_path):
     vision_tower = model.get_vision_tower()
     if not vision_tower.is_loaded:
         vision_tower.load_model()
-    vision_tower.to(device=model.device, dtype=torch.float16)
+    vision_tower.to(device=model.device, dtype=torch.float32)
     print("✅ The vision-tower is loaded successful!")
     # save
     model.save_pretrained(save_path)
